+++ packages/shared/src/index.ts
@@ -6,4 +6,5 @@
 export * from "./roles";
 export * from "./business-templates";
 export * from "./voice";
+export * from "./phone";
 export * from "./dto";